+++ web/env.html
@@ -147,10 +147,10 @@
           <a id="inspectButton" class="cw-button cyan" style="display: none;"
             ><i class="mdi mdi-18px mdi-magnify"></i>&nbsp; Inspect</a
           >
-          <a id="stopCompilationButton" class="cw-button red"
+          <a id="stopButton" class="cw-button red"
             ><i class="mdi mdi-18px mdi-stop"></i>&nbsp; Stop</a
           >
-          <a id="startCompilationButton" class="cw-button green"
+          <a id="startButton" class="cw-button green"
             ><i class="mdi mdi-18px mdi-play"></i>&nbsp; Run</a
           >
         </div>
