+++ web/js/doc.js
@@ -13,11 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-'use strict';
 
-import {
-  registerStandardHints
-} from './codeworld_shared.js'
+import { registerStandardHints } from './codeworld_shared.js';
 
 window.env = parent;
 const params = new URLSearchParams(window.location.search);
@@ -99,9 +96,9 @@ window.onscroll = (event) => {
           if (linkable && clickable) {
             pre.classList.add('clickable');
             pre.onclick = () => {
-              if (env && env.loadSample) {
+              if (env) {
                 savePosition();
-                env.loadSample(text);
+                env.postMessage({ type: 'loadSample', code: text }, '*');
               }
             };
           }
